Feat/Refresh Control 컴포넌트 삭제 후 개별 적용
