+++ screens/BottomTabs/home/generalBoard/PopularBoard.jsx
@@ -1,16 +1,32 @@
-import React from "react";
+import React, { useCallback, useState } from "react";
 import styled from "styled-components/native";
+import { RefreshControl } from "react-native";
 
 import { PostData } from "../../../../const/TempGeneralPostData";
 import PostPreview from "../../../../components/bottomTabs/home/board/PostPreview";
-import RefreshCtrl from "../../../../components/common/RefreshCtrl";
 
 const PopularBoard = () => {
+  // refresh 여부 state
+  const [refreshing, setRefreshing] = useState(false);
+
+  /* refresh될 때 호출되는 함수
+   * setRefreshing을 true로 만들었다가
+   * 2000ms 뒤에 다시 false로 변경
+   */
+  const onRefresh = useCallback(() => {
+    setRefreshing(true);
+    setTimeout(() => {
+      setRefreshing(false);
+    }, 2000);
+  }, []);
+
   return (
     <Container>
       <PostsWrap
         data={PostData}
-        refreshControl={<RefreshCtrl />}
+        refreshControl={
+          <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
+        }
         renderItem={({ item }) =>
           item.type === "popular" && <PostPreview postData={item} />
         }
